+++ CHANGELOG.md
@@ -7,6 +7,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- ## Unreleased -->
 
+## [0.4.14] 2019-11-04
+
 ### Added
 
 -   Added a `measurementExpression` option for overriding the default